move jwt guard into a guards folder since we will have more than one
